refactor: 전시 영역의 패키지 구조 리팩토링 진행 (#27)
