+++ src/main/java/com/hid_web/be/controller/exhibit/ExhibitController.java
@@ -4,7 +4,7 @@ import com.hid_web.be.controller.request.CreateExhibitRequest;
 import com.hid_web.be.controller.request.UpdateExhibitRequest;
 import com.hid_web.be.controller.response.ExhibitPreviewResponse;
 import com.hid_web.be.controller.response.ExhibitResponse;
-import com.hid_web.be.storage.ExhibitEntity;
+import com.hid_web.be.storage.exhibit.ExhibitEntity;
 import com.hid_web.be.domain.exhibit.ExhibitService;
 import lombok.AllArgsConstructor;
 import lombok.Data;
