+++ src/main/java/com/hid_web/be/controller/exhibit/request/CreateExhibitDetailRequest.java
