+++ src/main/java/com/hid_web/be/controller/exhibit/response/ExhibitPreviewResponse.java
@@ -1,6 +1,6 @@
 package com.hid_web.be.controller.response;
 
-import com.hid_web.be.storage.ExhibitEntity;
+import com.hid_web.be.storage.exhibit.ExhibitEntity;
 import lombok.*;
 
 @Getter
