+++ src/main/java/com/hid_web/be/storage/exhibit/ExhibitEntity.java
